requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.26.143
botocore==1.29.143
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
easyocr==1.7.2
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==1.0.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic_core==2.33.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.6.1
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.16.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tifffile==2025.5.10
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==1.26.20
uvicorn==0.23.2